simplified code for view all judges page /evaluate
