+++ components/buttons/DeleteButton.tsx
@@ -1,11 +1,11 @@
 import { FunctionComponent, useState } from "react";
 import styles from '../styles/Q.module.css';
 
-export const DeleteButton: FunctionComponent<{callback: Function, id: number, deleteMessage: string}> = ({callback, id, deleteMessage}) => {
+export const DeleteButton: FunctionComponent<{callback: Function, id: number | string, deleteMessage: string}> = ({callback, id, deleteMessage}) => {
   const [confirm, setConfirm] = useState(false);
   return (<div className={styles.deleteElement}>
 
-    <button onClick={(e) => {
+    <button onClick={(_) => {
       if (confirm) {
         callback(id);
       } else {
